Separate logic from collection configuration
